+++ changelog.md
@@ -8,9 +8,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Raise a label mismatch exception if label kwarg to Corpus constructor is inconsistent with automatically determined labels.
+- Test fixtures for Corpus creation
+- Test coverage for Corpus and Model creation
+
+### Changed
+- Update package dependencies versions.
 
 ### Fixed
 - `CorpusReader.train_batch_gen` raises StopIteration instead of returning None if no data can be generated.
+- Decoding from saved model is now possible for arbitrary Tensorflow model topologies that have the same input and output structure via named arguments that specify where input and output to the model occur.
+- RNN CTC model class now accepts `pathlib.Path` for directory argument
 
 ## [0.3.2]